extract all types to separate file
