+++ client/src/components/list/modal/ModalFooter.tsx
@@ -1,10 +1,6 @@
 import React from 'react';
 import modalFooter from '../../../styles/components/list/modal/modalFooter.styles';
-
-type ModalFooterProps = {
-  children: React.ReactNode;
-  className?: string;
-};
+import ModalFooterProps from '../../../types/components/list/modal/modalFooter.type';
 
 function ModalFooter({ children, className }: ModalFooterProps) {
   return <div className={`${modalFooter} ${className}`}>{children}</div>;
