+++ client/src/components/list/modal/ModalHeader.tsx
@@ -1,10 +1,6 @@
 import React from 'react';
 import modalHeader from '../../../styles/components/list/modal/modalHeader.styles';
-
-type ModalHeaderProps = {
-  children: React.ReactNode;
-  className?: string;
-};
+import ModalHeaderProps from '../../../types/components/list/modal/modalHeader.type';
 
 function ModalHeader({ children, className }: ModalHeaderProps) {
   return <div className={`${modalHeader} ${className}`}>{children}</div>;
