+++ client/src/types/components/list/journeys/addJourneyModal.type.ts
@@ -1,0 +1,6 @@
+type AddJourneyModalProps = {
+  open: boolean;
+  changeOpen: () => void;
+};
+
+export default AddJourneyModalProps;
