+++ client/src/types/components/list/listSection.type.ts
@@ -1,0 +1,5 @@
+type ListSectionProps = {
+  children: React.ReactNode;
+};
+
+export default ListSectionProps;
