+++ client/src/types/components/list/listTopBar.type.ts
@@ -1,0 +1,7 @@
+type ListTopBarProps = {
+  searchText: string;
+  setSearchText: (newSearchText: string) => void;
+  changeOpen: () => void;
+};
+
+export default ListTopBarProps;
