+++ client/src/types/components/list/modal/modal.type.ts
@@ -1,0 +1,8 @@
+type ModalProps = {
+  open: boolean;
+  changeOpen: () => void;
+  children: React.ReactNode;
+  className?: string;
+};
+
+export default ModalProps;
