+++ client/src/types/components/list/modal/modalBody.type.ts
@@ -1,0 +1,6 @@
+type ModalBodyProps = {
+  children: React.ReactNode;
+  className?: string;
+};
+
+export default ModalBodyProps;
