+++ client/src/types/components/list/modal/modalFooter.type.ts
@@ -1,0 +1,6 @@
+type ModalFooterProps = {
+  children: React.ReactNode;
+  className?: string;
+};
+
+export default ModalFooterProps;
