+++ client/src/types/components/list/modal/modalHeader.type.ts
@@ -1,0 +1,6 @@
+type ModalHeaderProps = {
+  children: React.ReactNode;
+  className?: string;
+};
+
+export default ModalHeaderProps;
