+++ client/src/types/components/list/stations/addStationModal.type.ts
@@ -1,0 +1,6 @@
+type AddStationModalProps = {
+  open: boolean;
+  changeOpen: () => void;
+};
+
+export default AddStationModalProps;
